# openapi/uacloudlibrary-api-1.0.1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UA Cloud Library REST Service",
    "description": "A REST-full interface to the CESMII & OPC Foundation Cloud Library",
    "contact": {
      "name": "OPC Foundation",
      "url": "https://opcfoundation.org/",
      "email": "office@opcfoundation.org"
    },
    "version": "1.0.1"
  },
  "paths": {
    "/infomodel/find": {
      "get": {
        "tags": [
          "InfoModel"
        ],
        "parameters": [
          {
            "name": "keywords",
            "in": "query",
            "description": "A list of keywords to search for in the information models. Specify * to return everything.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination limit",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Discovered OPC UA Information Model results of the models found in the UA Cloud Library matching the keywords provided.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UANodesetResult"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UANodesetResult"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UANodesetResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/infomodel/types/{identifier}": {
      "get": {
        "tags": [
          "InfoModel"
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "description": "OPC UA Information model identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The OPC UA Information model types.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The identifier provided could not be parsed.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The identifier provided could not be found.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/infomodel/download/{identifier}": {
      "get": {
        "tags": [
          "InfoModel"
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "description": "OPC UA Information model identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nodesetXMLOnly",
            "in": "query",
            "description": "Download NodeSet XML only, omitting metadata",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "metadataOnly",
            "in": "query",
            "description": "Download metadata only, omitting NodeSet XML",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The OPC UA Information model and its metadata.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UANameSpace"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UANameSpace"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UANameSpace"
                }
              }
            }
          },
          "400": {
            "description": "The identifier provided could not be parsed.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The identifier provided could not be found.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/infomodel/delete/{identifier}": {
      "get": {
        "tags": [
          "InfoModel"
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "description": "OPC UA Information model identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "forceDelete",
            "in": "query",
            "description": "Delete even if other nodesets depend on this nodeset.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The OPC UA Information model and its metadata.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UANameSpace"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UANameSpace"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UANameSpace"
                }
              }
            }
          },
          "400": {
            "description": "The identifier provided could not be parsed.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The identifier provided could not be found.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/infomodel/upload": {
      "put": {
        "tags": [
          "InfoModel"
        ],
        "parameters": [
          {
            "name": "overwrite",
            "in": "query",
            "description": "An optional flag if existing OPC UA Information models in the library should be overwritten.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UANameSpace"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UANameSpace"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UANameSpace"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UANameSpace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A status message indicating the successful upload.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "The provided nodeset file failed verification.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "An existing information model with the same identifier already exists in the UA Cloud Library and the overwrite flag was not set or the contributor name of existing information model is different to the one provided.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "The provided information model could not be stored or updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CloudLibRequiredModelInfo": {
        "type": "object",
        "properties": {
          "namespaceUri": {
            "type": "string",
            "nullable": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "availableModel": {
            "$ref": "#/components/schemas/Nodeset"
          }
        },
        "additionalProperties": false
      },
...
      "UANameSpace": {
        "required": [
          "category",
          "contributor",
          "copyrightText",
          "description",
          "license",
          "nodeset",
          "title"
        ],
        "type": "object",
        "properties": {
          "nodeset": {
            "$ref": "#/components/schemas/Nodeset"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "numberOfDownloads": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "contributor": {
            "$ref": "#/components/schemas/Organisation"
          },
          "license": {
            "minLength": 1,
            "type": "string"
          },
          "copyrightText": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "documentationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "iconUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "licenseUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "purchasingInformationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "releaseNotesUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "testSpecificationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "supportedLocales": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "additionalProperties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UAProperty"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UANodesetResult": {
        "required": [
          "category",
          "contributor",
          "copyrightText",
          "description",
          "license",
          "nodeset",
          "title"
        ],
        "type": "object",
        "properties": {
          "nodesetId": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "nodesetTitle": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "orgName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "version": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "publicationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "readOnly": true
          },
          "nodesetNamespaceUri": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "requiredNodesets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CloudLibRequiredModelInfo"
            },
            "nullable": true,
            "readOnly": true
          },
          "nodeset": {
            "$ref": "#/components/schemas/Nodeset"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "numberOfDownloads": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "minLength": 1,
            "type": "string"
          },
          "contributor": {
            "$ref": "#/components/schemas/Organisation"
          },
          "license": {
            "minLength": 1,
            "type": "string"
          },
          "copyrightText": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "documentationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "iconUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "licenseUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "purchasingInformationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "releaseNotesUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "testSpecificationUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "supportedLocales": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "additionalProperties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UAProperty"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UAProperty": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "basicAuth": [ ]
    },
    {
      "ApiKeyAuth": [ ]
    }
  ]
}